Youtube_automation/autotube-ai/backend/requirements.txt
# =============================================================================
# AutoTube AI — Python Dependencies
# =============================================================================

# --- Web Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
websockets==14.1

# --- Database ---
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
greenlet==3.1.1

# --- Task Queue ---
celery[redis]==5.4.0
redis==5.2.1

# --- AI Services ---
anthropic==0.42.0
elevenlabs==1.50.0
edge-tts==7.0.2
faster-whisper==1.1.0
# Stability AI accessed via httpx REST API (no grpcio dependency needed)

# --- Video Processing ---
moviepy==2.2.1
Pillow==11.1.0
pydub==0.25.1
numpy>=1.26.0

# --- YouTube API ---
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# --- HTTP & Scraping ---
httpx==0.28.1
beautifulsoup4==4.12.3
feedparser==6.0.11

# --- Config & Validation ---
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# --- Security ---
cryptography==44.0.0
python-jose[cryptography]==3.3.0

# --- Logging ---
loguru==0.7.3

# --- Retry Logic ---
tenacity==9.0.0

# --- Utilities ---
python-dateutil==2.9.0
aiofiles==24.1.0
